+++ src/Components/Like/Like.js
@@ -1,7 +1,6 @@
 import React from 'react'
 import {SideBar} from '../SideBar/SideBar'
-import { useAuth } from '../../Context/auth-context'
-import { useUser}  from '../../Context/user-context'
+import { useAuth  , useUser} from '../../Context'
 import VideoCard from '../Videos/VideoCard'
