+++ src/Components/PlayLists/PlayLists.js
@@ -1,8 +1,7 @@
 import React from 'react'
 import './PlayList.css'
 import {SideBar} from '../SideBar/SideBar'
-import { useAuth } from '../../Context/auth-context'
-import { useUser}  from '../../Context/user-context'
+import { useAuth , useUser} from '../../Context'
 import VideoCard from '../Videos/VideoCard'
 import { useNavigate } from 'react-router-dom'
 import {deletePlayList} from '../../Service/userAction'
