+++ src/Context/index.js
@@ -1,0 +1,3 @@
+export { useAuth , AuthProvider} from './auth-context'
+export {useVideos , VideosProvider} from './data-context'
+export {useUser , UserProvider} from './user-context'
